test(core): split advanced plugin tests
